0724 머지후 새작업
